chapter6/lesson17/requirements.txt
autogen-agentchat>=0.2.0
dashscope>=1.17.0
python-dotenv>=1.0.0 